requirements.txt
psycopg[binary]>=3.0.0
python-dotenv>=0.19.0
openai>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0